+++ src/main/java/me/jasperchasetoq/builderskit/implementations/items/BuildersBlock.java
@@ -1,0 +1,28 @@
+package me.jasperchasetoq.builderskit.implementations.items;
+
+import io.github.thebusybiscuit.slimefun4.api.events.PlayerRightClickEvent;
+import io.github.thebusybiscuit.slimefun4.api.items.ItemGroup;
+import io.github.thebusybiscuit.slimefun4.api.items.SlimefunItem;
+import io.github.thebusybiscuit.slimefun4.api.items.SlimefunItemStack;
+import io.github.thebusybiscuit.slimefun4.api.recipes.RecipeType;
+import io.github.thebusybiscuit.slimefun4.core.handlers.BlockUseHandler;
+import org.bukkit.inventory.ItemStack;
+
+public class BuildersBlock extends SlimefunItem {
+
+    public BuildersBlock(ItemGroup itemGroup, SlimefunItemStack item, RecipeType recipeType, ItemStack[] recipe) {
+        super(itemGroup, item, recipeType, recipe);
+    }
+
+    @Override
+    public void preRegister() {
+        BlockUseHandler blockUseHandler = this::onBlockRightClick;
+        addItemHandler(blockUseHandler);
+    }
+
+    private void onBlockRightClick(PlayerRightClickEvent event) {
+
+    }
+
+
+}
