+++ src/main/java/me/jasperchasetoq/builderskit/utils/Setup.java
@@ -1,0 +1,11 @@
+package me.jasperchasetoq.builderskit.utils;
+
+import me.jasperchasetoq.builderskit.BuildersKit;
+import me.jasperchasetoq.builderskit.implementations.BuilderItems;
+
+public class Setup {
+    public static void setup(BuildersKit bk) {
+        BuilderItems.setup(bk);
+        Category.setup(bk);
+    }
+}
